requirements.txt
appdirs==1.4.3
bcrypt==3.1.4
cfe==0.0.15
dj-database-url==0.4.2
Django==2.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.0
djangorestframework==3.7.7
gunicorn==19.7.1
olefile==0.45.1
packaging==16.8
Pillow==5.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
pytz==2018.3
six==1.11.0
whitenoise==3.3.1
